+++ src/pages/myPage/components/myPost/MyPostsCard.tsx
