+++ src/pages/myPage/components/profile/InformationCard.tsx
